requirements.txt
fastapi==0.66.0
SQLAlchemy==1.4.20
pydantic==1.8.2
psycopg2-binary==2.9.1
alembic==1.6.5
uvicorn==0.14.0
asyncpg==0.24.0
requests
firebase-admin
fastapi_pagination
pyjwt
httpx
orjson
vine==1.3.0
celery==4.3.0
celery-pool-asyncio==0.2.0
redis==3.5.3